requirements.txt
bs4
faiss-cpu
Flask
google-ai-generativelanguage
google-generativeai
langchain-community
pandas
pymongo
sentence-transformers
yfinance
